requirements.txt
scikit-learn>=0.23.1
numpy>=1.18.5
matplotlib>=3.2.2
scipy>=1.5.0
pyyaml>=5.4.1

eif~=2.0.2
opylib~=0.9.0